% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{city_module_ui}
\alias{city_module_ui}
\title{Shiny Module UI for City Weather and Ice Cream Estimation}
\usage{
city_module_ui(id)
}
\arguments{
\item{id}{Module ID for Shiny namespace.}
}
\description{
Shiny Module UI for City Weather and Ice Cream Estimation
}
