% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cone_radius}
\alias{cone_radius}
\title{Calculate cone radius based on input x}
\usage{
cone_radius(x)
}
\arguments{
\item{x}{Numeric value indicating the position along the cone's height.}
}
\value{
Numeric value representing the radius of the cone at position x.
}
\description{
Calculate cone radius based on input x
}
\examples{
cone_radius(5)
cone_radius(9)
cone_radius(11)
}
