% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cone_radius_derivative}
\alias{cone_radius_derivative}
\title{Derivative of Cone Radius}
\usage{
cone_radius_derivative(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector representing the derivative of cone radius.
}
\description{
Function to compute the derivative of the cone radius
}
\examples{
cone_radius_derivative(c(5, 0))
}
