% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cone_radius_for}
\alias{cone_radius_for}
\title{Vectorized function for cone_radius using for loop}
\usage{
cone_radius_for(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector with calculated radii for each x.
}
\description{
Vectorized function for cone_radius using for loop
}
