% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cone_radius_squared}
\alias{cone_radius_squared}
\title{Square of Cone Radius}
\usage{
cone_radius_squared(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of squared radii.
}
\description{
Function to compute the square of the cone radius
}
\examples{
cone_radius_squared(c(0, 5))
}
\author{
Ronald Medvedev
}
