% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_app}
\alias{run_app}
\title{Run Shiny App for Weather and Ice Cream Simulation}
\usage{
run_app(api_key)
}
\arguments{
\item{api_key}{Character string of OpenWeatherMap API key.}
}
\description{
Run Shiny app for weather and ice cream simulation
}
\examples{
\dontrun{run_app("your_api_key")}
}
