% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_city_validity}
\alias{check_city_validity}
\title{Check City Validity via API}
\usage{
check_city_validity(city_name, api_key)
}
\arguments{
\item{city_name}{Character string representing the name of the city to check.}

\item{api_key}{Character string of OpenWeatherMap API key.}
}
\value{
Logical value indicating if the city is valid.
}
\description{
Check if a city name is valid by querying the OpenWeatherMap API
}
\examples{
\dontrun{check_city_validity("Lausanne", "your_api_key")}
}
