% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{city_module_server}
\alias{city_module_server}
\title{Shiny Module Server for City Weather and Ice Cream Estimation}
\usage{
city_module_server(id, api_key)
}
\arguments{
\item{id}{Module ID for Shiny namespace.}

\item{api_key}{OpenWeatherMap API key.}
}
\description{
Shiny Module Server for City Weather and Ice Cream Estimation
}
