% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cone_radius_vectorize}
\alias{cone_radius_vectorize}
\title{Vectorized function for cone_radius using Vectorize}
\usage{
cone_radius_vectorize(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector with calculated radii for each x.
}
\description{
Vectorized function for cone_radius using Vectorize
}
