% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_forecast}
\alias{get_forecast}
\title{Retrieve the 5-day weather forecast for a specified city}
\usage{
get_forecast(city, api_key, units = "metric")
}
\arguments{
\item{city}{Character string of the city name to get the forecast for.}

\item{api_key}{Character string of the OpenWeatherMap API key.}

\item{units}{Character string specifying the units (default is "metric").}
}
\value{
Data frame containing the forecast data if successful, NULL otherwise.
}
\description{
Retrieve the 5-day weather forecast for a specified city
}
