% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_app}
\alias{run_app}
\title{Run Shiny app for weather and ice cream simulation}
\usage{
run_app(api_key)
}
\arguments{
\item{api_key}{Character string of OpenWeatherMap API key.}
}
\description{
This function launches a Shiny app that allows the user to check the weather forecast
for a city and run a simulation for ice cream consumption.
}
